Add automated plural suffix addition for title
This commit adds a macro to automatically add the plural "s" suffix to
the "Author" and "Supervisor" words in the title page based on the
number of supplied authors and supervisors.

This commit also Closes #14 by removing the accidentally hardcoded
supervisor name which did not use the one supplied by the \supervisor
macro.

Summary of Changes:
- Add makeplural macro to automatically add plural suffixes
- Fix bug caused by hardcoded supervisor name