Simplify declaration of formatted bibliography
This commit addresses Issue [#13] by simplifying the commands required
to generate a formatted bibliography that is included in the Table of
Contents. This is accomplished by redefining the font used for the
bibliography to use the \raggedright option.

Since the bibliography can now be generated by a single
\printbibliography command, the \BibliographyName macro is no longer
required. However, it is left in the tud-report.cls file so that users
still trying to compile the report with the old invocation will not get
failing builds.

[13]: https://github.com/skilkis/tudelft-light/issues/13